servers/auth/pyproject.toml
[tool.poetry]
name = "auth-dtaas"
version = "0.1.0"
description = "Auth Microservice (DTaaS) Package"
authors = ["Asti <f20190640@goa.bits-pilani.ac.in>"]
readme = "README.md"
packages = [
    {include = "auth_dtaas"}
]

[tool.poetry.dependencies]
python = "^3.10"
requests-oauthlib = "^1.3.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
